GSYE-388: Move Simulation classes into gsy-e-core.simulation.__init__ and rename packages in there
