+++ src/gsy_e/gsy_e_core/simulation/results_manager.py
@@ -33,8 +33,8 @@ from gsy_e.models.area.scm_manager import SCMManager
 
 if TYPE_CHECKING:
     from gsy_e.models.area import Area, AreaBase, CoefficientArea
-    from simulation_setup import SimulationSetup
-    from gsy_e.gsy_e_core.simulation_utils.simulation_progress_info import SimulationProgressInfo
+    from setup import SimulationSetup
+    from gsy_e.gsy_e_core.simulation.progress_info import SimulationProgressInfo
 
 log = getLogger(__name__)
