+++ src/gsy_e/gsy_e_core/simulation/status_manager.py
